refactor(geode-util): migrate JSON Fixture loader + schema + golden accessors
Phase 2.1 of Epic #429. Move the JSON `Fixture` machinery from
`geode-validation/src/fixture.rs` into `geode_util::fixture`, with
`geode-validation` re-exporting so the reference-test suite keeps its
existing `geode_validation::Fixture` / `fixture.output_f64(...)` call paths.

Moved to `geode_util::fixture`:
- Schema types: `Fixture`, `Field`, `OutputField`, `Provenance`,
  `FixtureError`, `FixtureFormat`, `SUPPORTED_SCHEMA_VERSIONS`.
- Loaders: `load_from`, `load`, `load_json` (now calls
  `geode_util::repo::fixture_path` directly, no cross-crate hop).
- Golden accessors: `output_f64`/`GoldenF64`, `output_c128`/`GoldenC128`,
  `input_c128`, `iter_outputs`, and `pub(crate) flatten_to_f64`.
  `input_c128`/`output_c128` keep using `geode_util::interop`
  (now an in-crate `crate::interop` call).

Added `thiserror` to geode-util for the migrated `FixtureError` derive.

Stays in geode-validation:
- `diff.rs` + `ComparisonReport`/`FieldDiff` (the validation diff artifact).
- The comparison entry points. Because `Fixture` is now a foreign type, the
  orphan rule forbids `compare_against`/`compare_complex_against` as inherent
  methods; they become `geode-validation` free functions wrapping
  `diff::compare`/`diff::compare_complex`, re-exported at the crate root.
  The 28 reference-test call sites are repointed from
  `fixture.compare_against(&actual)` to
  `geode_validation::compare_against(&fixture, &actual)` (and the complex
  variant).

geode-core does not depend on geode-util. `cargo build` workspace,
`cargo clippy -p geode-util -p geode-validation --all-targets -- -D warnings`,
and `cargo test -p geode-util` / `-p geode-validation` all pass.
