+++ crates/geode-validation/tests/sphere_mie_jax_reference.rs
@@ -491,7 +491,7 @@ fn jax_mie_small_spectrum_agrees_with_burn() {
         burn_physical.clone(),
     );
 
-    let report = fixture.compare_complex_against(&actual);
+    let report = geode_validation::compare_complex_against(&fixture, &actual);
     if !report.passed {
         eprintln!(
             "sphere_mie_small (JAX) complex-comparator report: {:#?}",
